backend/requirements.txt
Flask
Flask-Cors
requests
google-generativeai
python-dotenv
Werkzeug
gunicorn

